refactor: mask files (#292)
