+++ src/mask/hooks/usePostInfo.ts
@@ -1,7 +1,5 @@
-'use client';
-
 import { DOMProxy } from '@dimensiondev/holoflows-kit';
-import { type PostContext, PostInfoProvider } from '@masknet/plugin-infra/content-script';
+import { type PostContext } from '@masknet/plugin-infra/content-script';
 import {
     createConstantSubscription,
     EMPTY_ARRAY,
@@ -13,22 +11,15 @@ import {
 } from '@masknet/shared-base';
 import { makeTypedMessageEmpty, makeTypedMessageTuple } from '@masknet/typed-message';
 import { compact } from 'lodash-es';
-import { memo, type PropsWithChildren, useMemo } from 'react';
+import { useMemo } from 'react';
 import urlcat from 'urlcat';
 
 import { SITE_HOSTNAME, SITE_URL } from '@/constants/index.js';
-import type { EncryptedPayload } from '@/helpers/getEncryptedPayload.js';
 import { getPostUrl } from '@/helpers/getPostUrl.js';
-import { DecryptMessage } from '@/mask/main/DecryptMessage.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
 
-interface Props extends PropsWithChildren<{}> {
-    post: Post;
-    payload: EncryptedPayload;
-}
-
-export const DecryptPost = memo(function DecryptPost({ post, payload, children }: Props) {
-    const postInfo = useMemo((): PostContext => {
+export function usePostInfo(post: Post) {
+    return useMemo((): PostContext => {
         const author = ProfileIdentifier.of(SITE_HOSTNAME, post.author.displayName).unwrapOr(null);
         const imageUris: string[] = compact(
             post.metadata.content?.attachments
@@ -69,18 +60,4 @@ export const DecryptPost = memo(function DecryptPost({ post, payload, children }
             },
         };
     }, [post]);
-
-    return (
-        <PostInfoProvider post={postInfo}>
-            {children}
-            <div
-                onClick={(e) => {
-                    e.preventDefault();
-                    e.stopPropagation();
-                }}
-            >
-                <DecryptMessage text={payload[0]} version={payload[1]} />
-            </div>
-        </PostInfoProvider>
-    );
-});
+}
