+++ src/mask/main/decrypt.ts
