+++ src/mask/main/registry.ts
